Seperate into exec and library
